+++ Runtime/Scripts/GltfGlobals.cs
@@ -29,23 +29,26 @@ namespace GLTFast {
         BIN = 0x004e4942
     }
 
-    static class GltfGlobals {
+    /// <summary>
+    /// Generic glTF constants and utility methods.
+    /// </summary>
+    public static class GltfGlobals {
         
         /// <summary>
         /// glTF-Binary file extension
         /// </summary>
-        internal const string glbExt = ".glb";
+        public const string glbExt = ".glb";
         
         /// <summary>
         /// glTF file extension
         /// </summary>
-        internal const string gltfExt = ".gltf";
+        public const string gltfExt = ".gltf";
         
         /// <summary>
         /// First four bytes of a glTF-Binary file are made up of this signature
         /// Represents glTF in ASCII
         /// </summary>
-        internal const uint GLB_MAGIC = 0x46546c67; 
+        public const uint GLB_MAGIC = 0x46546c67; 
         
         /// <summary>
         /// Figures out if a byte array contains data of a glTF-Binary